Fix jumpy scroll behaviour (#607)
* Fix jumpy scroll behaviour

I'm not sure if this is the best fix for the issue (I would have liked it to allow the browser to deal with moving the cursor into view.) but for now I think this is better than what it was before. I tried to imitate what chrome was doing when the cursor was out of view. (Except when it was partially in view but I decided to just keep it simple) 

This basically checks the cursor's position and if it isn't in view it'll move the cursor into the middle of the view. 

Fixes #279
Fixes #589

* Subtract Height